- made it so you cannot delete/change the role of your own account - moved user exceptions and advice to their own folder

+++ src/main/java/edu/ntt/hotelbookingplatform/controller/UserController.java
@@ -6,6 +6,7 @@ import edu.ntt.hotelbookingplatform.dto.in.ChangeRoleDTO;
 import edu.ntt.hotelbookingplatform.dto.in.UserCreationDTO;
 import edu.ntt.hotelbookingplatform.dto.mapper.UserMapper;
 import edu.ntt.hotelbookingplatform.dto.out.UserDTO;
+import edu.ntt.hotelbookingplatform.exception.user.CannotChooseOwnUserException;
 import edu.ntt.hotelbookingplatform.service.UserService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
@@ -48,8 +49,12 @@ public class UserController {
     }
 
     @PatchMapping("/changeRole")
-    public ResponseEntity<String> changeUserRole(@Valid @RequestBody ChangeRoleDTO changeRoleDTO){
-        userService.updateUserRole(changeRoleDTO.getEmail(), changeRoleDTO.getRole());
+    public ResponseEntity<String> changeUserRole(@Valid @RequestBody ChangeRoleDTO changeRoleDTO, @AuthenticationPrincipal String currentUserEmail ){
+        String email = changeRoleDTO.getEmail();
+        String role = changeRoleDTO.getRole();
+        if(currentUserEmail.equals(email))
+            throw new CannotChooseOwnUserException();
+        userService.updateUserRole(email, role);
         return ResponseEntity.ok("User role changed successfully!");
     }
 
@@ -66,7 +71,9 @@ public class UserController {
     }
 
     @DeleteMapping("/{email}")
-    public ResponseEntity<String> deleteUser(@PathVariable @Email(message="Email must have a valid format!")  String email){
+    public ResponseEntity<String> deleteUser(@PathVariable @Email(message="Email must have a valid format!")  String email, @AuthenticationPrincipal String currentUserEmail){
+        if(currentUserEmail.equals(email))
+            throw new CannotChooseOwnUserException();
         userService.deleteUser(email);
         return ResponseEntity.ok("User account deleted successfully!");
     }
